wiso: resolve a specific key to a curated list
Whenever a package would contain the string "wiso UB Leipzig Profil",
replace the whole list of packages with a list of curated WISO database
names to use.

This can work, because every record will carry its database name in the
package field as well:

    $ taskcat GeniosCombinedIntermediateSchema | head -1 | jq '.["x.packages"]'
    [
      "IJAR",
      "Genios (Sozialwissenschaften)",
      "Genios (Fachzeitschriften)",
      "journal_sozi"
    ]

This is a mere workaround and should be removed for a simpler approach.