dbinet: refs #13653
Workaround for dead links and silly dates.

Most silly dates are more abtract things, like websites, studies, themes and so
on - things that do not have temporal restrictions, in German: zeitlich
übergreifend.

...
<Category facetteID="53" facetteLabel="temporalcoverage" handle="25903" id="6914" term="XXX">
    <Name handle="25897" xml:lang="de">[06914]zeitlich übergreifend / unabhängig</Name>
    <Description handle="25899" xml:lang="de">zeitlich übergreifend / unabhängig&gt;&gt;</Description>
</Category>
...

If a link is reachable (checked via clinker, https://git.io/fAC27), we assume
it is valid NOW, and put in the current date and set the type to
"ElectronicIntegratingResource".

Workaround is open to debate. The format seems to be DBClear [1].

[1] http://fiz1.fh-potsdam.de/volltext/iuk03/08019.pdf