apply automatic style across files, refs #14779
The command executed was:

    make imports style

It applies a) import sorting with isort [1] and b) basic PEP8 [2] styles
with yapf [3].  The goal is to try out a consistent style and see it we
can converge. If we can, we win readability.

Two libraries are required for this to work: isort and yapf.

> Random code style discussions: https://stackoverflow.com/a/39874565

The application of local helpers such as git hooks it at the discretion
of the developer.

[1] https://github.com/timothycrosley/isort
[2] https://www.python.org/dev/peps/pep-0008/
[3] https://github.com/google/yapf