Parameterize python executable, refs #15882
For the moment, many transformation tasks run as subprocesses, using
just "python" as executable. On many distributions, this will already by
Python 3, but not on all.

Allow to set the python executable path in the siskin.ini, e.g:

    [core]

    ...
    python = python3.6
    ...

Note: There is no default used at the moment, so this entry in the
config is mandatory (even if it is just a "python = python").