ai: introduce --filter-config-style flag for AIUpdate
currently two values:

* default: current setup
* reduced: only take into account holding files for institutions

The reduced style may help us to work around ever changing collection
lists in e.g. crossref which would have to be maintained in AMSL as
well. We need to test this first - so we add a flag so we can generate
both versions side by side.